Put back QR code address scanning and enforce scan security
This commit prevent scanning non-EIP addresses, or more generally
prevent scanning address that generate warning for recipient (currently
the only possible warning is for non-EIP address for ETH/ETC).

If the warning-address is entered with keyboard instead of scanning, it
will still display the warning (as now), the behaviour doesnt change.

Closes #1412
Closes #1406