Update Babel config to prepare working cli
Directly registering babel as it is was triggering cryptic errors, like:

    Error: Reentrant preset detected trying to load "<path>/babel.config.js".
    This module is not ignored and is trying to load itself while compiling
    itself, leading to a dependency cycle.

This commits moves babel to the last version, so as it go > than 7-beta.55,
we have to get rid of stage-0 and replace it by corresponding specific
presets (I think we can clean a lot of things from here).

It also involve the `process.env.CLI` var, to use correct babel
transform on modules.