Rollback to previous electron and electron-updater
The 1.8.0 introduced a regression on Mac regarding bad implementation of @ledgerhq/electron-updater
The quick fixes for now is to rollback this and reimplement a better version later

The issue found was that: when upgrading from 1.8.0 to 1.8.1, the 'Install Now' button doesn't do anything, it should have quitAndInstall (it works on Linux and Windows). Also there is race conditions if you click multiple times on quitAndInstall so we're going to wait for a better implementation (fixes to be tested https://github.com/LedgerHQ/electron-builder/pull/1 ).