LL-1556 Account migration flow
Adds the migration flow when necessary. It will show a banner on
the portfolio screen if there are eligible accounts to be migrated.
On this modal flow we will iterate all different cryptocurrencies
and after each loop we will suggest the user changing devices until
the migration is completed.

+-----------+   +-------------+   +-------------+   +-------------+
|           |   |             |   |             |   |             |
| Overview  ---->   Connect   ---->   Migrate   ----->  Summary   |
|           |   |             |   |             |   |             |
+-----^-----+   +------^------+   +-------------+   +------|------+
      |                |                                   v
      |                |                              Looped all
      |                +-------------------------no---currencies?
    yes                                                    |
      |                                                    |
      +------More migratable------------------------------yes
             accounts?
                |
               no
                |
                v
           +---------+
           |  DONE   |
           +---------+

Feature

https://ledgerhq.atlassian.net/browse/LL-1556

To test this you will need ideally two devices with both ETH and ETC
that you imported with the experimental libcore OFF. Then you can
enable experimental libcore flag and the banner should appear on top