Swap - Cleanup the account selector state in Form/index.js
There was some logic for filtering available accounts at the form level, however this was also handled inside the logic of the currency account selector, resulting in a mishmash of state and doubled useless code. Removing all the logic from the Form/index.js that handled accounts and allowing the child From/To to communicate the account changes seems to fix a couple of bugs as well as simplify a bunch of code.

Relying on the the accounts returned by internal logic of the selector component makes more sense than that of a parent component that should be agnostic to this knowledge.