LL-5947 Change the TC/PP screen into a Legal screen (#3967)
* chore(i18n): update terms translations

- Update *label keys
- Merge switch* keys to only one key as designed.
- Change title as requested by design

* chore(config): add support to dynamic terms urls

Thanks to this new hook, url to terms webpage displayed is now based on
user language. Terms supports `en`, `fr`, `es` and `de`. `en` is used as
the fallback language.
- Remove TERMS*.md files
- Remove unused functions used until now to dynamically load TERMS*.md files

* design(onboarding): redesign terms modal as a page

- Create new terms page to replace old terms modal.
- Remove terms modal and all related file only used by the onboarding process.
- Remove CheckTermsAccepted component used so far by the SelectDevice step
to conditionnaly display the terms modal.

* design(settings): add link to external terms page

Until now, terms content was loaded from a markdown file and displayed to
the user in a modal.
From now, user will be redirected to a terms page hosted in our website.
Based on current user language, the user will be redirected to the correct
terms page.
- Remove TermsButton component

* chore(terms): generic hooks to load dynmamic urls

This hook will be used to load dynamic urls based on current user language.

* test: update snapshot