feat(account): improve the empty state
Until now, the header in the account page for an empty account (no txs) wasn't
rendered. Because of this, walletconnect button wasn't rendered, you needed
some tokens to see it appears. This commit fixes this, from now, walletconnect
button and receive/buy buttons would be rendered for any empty accounts.
As it is not required to own tokens to use walletconnect, this modification makes
sense to me. Regarding the receive/buy buttons, I think it's better to have them
always available, even for empty accounts (we missed a business opportunity by
hidding the buy button in this case).